refactor start options
