+++ packages/expo-cli/src/commands/start/parseStartOptions.ts
@@ -1,0 +1,141 @@
+import { ExpoConfig, isLegacyImportsEnabled } from '@expo/config';
+import { Project, ProjectSettings } from '@expo/xdl';
+
+import Log from '../../log';
+import { URLOptions } from '../../urlOpts';
+
+export type NormalizedOptions = URLOptions & {
+  webOnly?: boolean;
+  dev?: boolean;
+  minify?: boolean;
+  https?: boolean;
+  nonInteractive?: boolean;
+  clear?: boolean;
+  maxWorkers?: number;
+  sendTo?: string;
+  host?: string;
+  lan?: boolean;
+  localhost?: boolean;
+  tunnel?: boolean;
+};
+
+export type Options = NormalizedOptions & {
+  parent?: { nonInteractive: boolean; rawArgs: string[] };
+};
+
+function hasBooleanArg(rawArgs: string[], argName: string): boolean {
+  return rawArgs.includes('--' + argName) || rawArgs.includes('--no-' + argName);
+}
+
+function getBooleanArg(rawArgs: string[], argName: string): boolean {
+  if (rawArgs.includes('--' + argName)) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+// The main purpose of this function is to take existing options object and
+// support boolean args with as defined in the hasBooleanArg and getBooleanArg
+// functions.
+export async function normalizeOptionsAsync(
+  projectDir: string,
+  options: Options
+): Promise<NormalizedOptions> {
+  const opts: NormalizedOptions = {
+    ...options, // This is necessary to ensure we don't drop any options
+    webOnly: !!options.webOnly, // This is only ever true in the start:web command
+    nonInteractive: options.parent?.nonInteractive,
+  };
+
+  const rawArgs = options.parent?.rawArgs || [];
+
+  if (hasBooleanArg(rawArgs, 'dev')) {
+    opts.dev = getBooleanArg(rawArgs, 'dev');
+  } else {
+    opts.dev = true;
+  }
+  if (hasBooleanArg(rawArgs, 'minify')) {
+    opts.minify = getBooleanArg(rawArgs, 'minify');
+  } else {
+    opts.minify = false;
+  }
+  if (hasBooleanArg(rawArgs, 'https')) {
+    opts.https = getBooleanArg(rawArgs, 'https');
+  } else {
+    opts.https = false;
+  }
+
+  if (hasBooleanArg(rawArgs, 'android')) {
+    opts.android = getBooleanArg(rawArgs, 'android');
+  }
+
+  if (hasBooleanArg(rawArgs, 'ios')) {
+    opts.ios = getBooleanArg(rawArgs, 'ios');
+  }
+
+  if (hasBooleanArg(rawArgs, 'web')) {
+    opts.web = getBooleanArg(rawArgs, 'web');
+  }
+
+  if (hasBooleanArg(rawArgs, 'localhost')) {
+    opts.localhost = getBooleanArg(rawArgs, 'localhost');
+  }
+
+  if (hasBooleanArg(rawArgs, 'lan')) {
+    opts.lan = getBooleanArg(rawArgs, 'lan');
+  }
+
+  if (hasBooleanArg(rawArgs, 'tunnel')) {
+    opts.tunnel = getBooleanArg(rawArgs, 'tunnel');
+  }
+
+  await cacheOptionsAsync(projectDir, opts);
+  return opts;
+}
+
+async function cacheOptionsAsync(projectDir: string, options: NormalizedOptions): Promise<void> {
+  await ProjectSettings.setAsync(projectDir, {
+    devClient: options.devClient,
+    scheme: options.scheme,
+    dev: options.dev,
+    minify: options.minify,
+    https: options.https,
+  });
+}
+
+export function parseStartOptions(
+  options: NormalizedOptions,
+  exp: ExpoConfig
+): Project.StartOptions {
+  const startOpts: Project.StartOptions = {};
+
+  if (options.clear) {
+    startOpts.reset = true;
+  }
+
+  if (options.nonInteractive) {
+    startOpts.nonInteractive = true;
+  }
+
+  if (options.webOnly) {
+    startOpts.webOnly = true;
+  }
+
+  if (options.maxWorkers) {
+    startOpts.maxWorkers = options.maxWorkers;
+  }
+
+  if (options.devClient) {
+    startOpts.devClient = true;
+  }
+
+  if (isLegacyImportsEnabled(exp)) {
+    // For `expo start`, the default target is 'managed', for both managed *and* bare apps.
+    // See: https://docs.expo.io/bare/using-expo-client
+    startOpts.target = options.devClient ? 'bare' : 'managed';
+    Log.debug('Using target: ', startOpts.target);
+  }
+
+  return startOpts;
+}
